Update exam.Rmd
Almost done on my part.  The only thing remaining on my part is to ensure that all my variables are coded as factors and not integers.  I also still need to delete all the "View" statements and stuff like that, which is not necessary and makes the code needlessly long.